refactor: moved getDimensions to helper directory
